requirements.txt
boto3==1.28.2
elasticsearch==8.8.2
jinjasql==0.1.8
JSONBender==0.9.3
jsonschema==4.18.3
neo4j==5.10.0
numpy==1.25.1
pandas==1.4.2
psycopg2==2.9.6
pymongo==4.4.0
PyYAML==6.0
PyYAML==6.0
snowflake_connector_python==3.0.4
#MySQL-python==1.2.5
